replay buffer sampling improving; trying tf weights

+++ replay_buffers/replay_buffers.py
@@ -139,7 +139,7 @@ class PrioritizedReplayBuffer(ReplayBuffer):
         it_min = self._it_min.min()
         p_min = it_min / it_sum
         max_weight = (p_min * len(self._storage)) ** (-self._beta)
-        p_sample = tf.stack([self._it_sum[idx] / it_sum for idx in idxes])
+        p_sample = tf.stack([self._it_sum[idx] / it_sum for idx in idxes], dtype='float32')
         weights = (p_sample*len(self._storage)) ** (-self._beta)
         weights = weights / max_weight
         encoded_sample = self._encode_sample(idxes)
